requirements.txt
requests>=2.32.3
setuptools>=80.3.0
pyyaml